wrote a single object property to the page (looped)
Firgured out how to get the loop to write to the page instead of just the console! Was using 'textContent' instead of 'innerHTML' when trying to 'appendChild' a new P element onto a DIV (textContent doesn't create a Node, so appendChild doesn't have a valid reference of where it can attach.